Move Hibernate Search DevUI tests to hibernate-search-orm-elasticsearch/deployment
Because we'll need to test the Standalone extension too, and we can't
keep adding dependencies to a single integration test module.

(cherry picked from commit 3b290d24bde844aba711ebf005ef1416913501aa)
