+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/orm/elasticsearch/test/devui/DevUISmokeTest.java
@@ -1,0 +1,18 @@
+package io.quarkus.hibernate.search.orm.elasticsearch.test.devui;
+
+import org.junit.jupiter.api.extension.RegisterExtension;
+
+import io.quarkus.test.QuarkusDevModeTest;
+
+public class DevUISmokeTest extends AbstractDevUITest {
+
+    @RegisterExtension
+    static final QuarkusDevModeTest test = new QuarkusDevModeTest()
+            .withApplicationRoot(
+                    (jar) -> jar.addAsResource("application-devui.properties", "application.properties")
+                            .addClasses(MyIndexedEntity.class));
+
+    public DevUISmokeTest() {
+        super("<default>", MyIndexedEntity.class.getName());
+    }
+}
