+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/orm/elasticsearch/test/devui/MyIndexedEntity.java
@@ -1,4 +1,4 @@
-package io.quarkus.test.devui.namedpu;
+package io.quarkus.hibernate.search.orm.elasticsearch.test.devui;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -8,7 +8,7 @@ import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 
 @Entity
 @Indexed
-public class MyNamedPuIndexedEntity {
+public class MyIndexedEntity {
 
     @Id
     Long id;
