+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/java/io/quarkus/hibernate/search/orm/elasticsearch/test/devui/namedpu/MyNamedPuIndexedEntity.java
@@ -1,4 +1,4 @@
-package io.quarkus.test.devui;
+package io.quarkus.hibernate.search.orm.elasticsearch.test.devui.namedpu;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -8,7 +8,7 @@ import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
 
 @Entity
 @Indexed
-public class MyIndexedEntity {
+public class MyNamedPuIndexedEntity {
 
     @Id
     Long id;
