+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/resources/application-devui-active-false-and-named-pu-active-true.properties
@@ -1,0 +1,27 @@
+# We already start Elasticsearch with Maven
+quarkus.devservices.enabled=false
+
+quarkus.datasource.db-kind=h2
+quarkus.datasource.jdbc.url=jdbc:h2:mem:default;DB_CLOSE_DELAY=-1
+
+quarkus.datasource."nameddatasource".db-kind=h2
+quarkus.datasource."nameddatasource".jdbc.url=jdbc:h2:mem:nameddatasource;DB_CLOSE_DELAY=-1
+
+quarkus.hibernate-orm.datasource=<default>
+quarkus.hibernate-orm.packages=io.quarkus.hibernate.search.orm.elasticsearch.test.devui
+quarkus.hibernate-orm.database.generation=drop-and-create
+
+quarkus.hibernate-orm."namedpu".datasource=nameddatasource
+quarkus.hibernate-orm."namedpu".packages=io.quarkus.hibernate.search.orm.elasticsearch.test.devui.namedpu
+quarkus.hibernate-orm."namedpu".database.generation=drop-and-create
+
+# Hibernate Search is inactive for the default PU
+quarkus.hibernate-search-orm.active=false
+quarkus.hibernate-search-orm.elasticsearch.version=8.12
+quarkus.hibernate-search-orm.elasticsearch.hosts=${elasticsearch.hosts:localhost:9200}
+quarkus.hibernate-search-orm.schema-management.strategy=drop-and-create-and-drop
+
+# ... but it's (implicitly) active for a named PU
+quarkus.hibernate-search-orm."namedpu".elasticsearch.version=8.12
+quarkus.hibernate-search-orm."namedpu".elasticsearch.hosts=${elasticsearch.hosts:localhost:9200}
+quarkus.hibernate-search-orm."namedpu".schema-management.strategy=drop-and-create-and-drop
