+++ extensions/hibernate-search-orm-elasticsearch/deployment/src/test/resources/application-devui-active-false.properties
@@ -1,0 +1,12 @@
+# We already start Elasticsearch with Maven
+quarkus.devservices.enabled=false
+
+quarkus.datasource.db-kind=h2
+quarkus.datasource.jdbc.url=jdbc:h2:mem:default;DB_CLOSE_DELAY=-1
+
+quarkus.hibernate-orm.database.generation=drop-and-create
+
+quarkus.hibernate-search-orm.active=false
+quarkus.hibernate-search-orm.elasticsearch.version=8.12
+quarkus.hibernate-search-orm.elasticsearch.hosts=${elasticsearch.hosts:localhost:9200}
+quarkus.hibernate-search-orm.schema-management.strategy=drop-and-create-and-drop
